refactor: Rectangle의 부모 클래스 BaseShape 완성
- BaseShape는 Colorable을 채택하지 않음
    - Photo에서 Colorable을 사용하지 않기 때문에 불필요한 프로퍼티/메소드를 없애고자 채택 안함
- 따라서 Rectangle에서 Colorable 채택
