+++ Drawing-iPad/Source/Extensions/NotificationCenter+Name.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 extension Notification.Name {
-    static let rectangleUpdated = Notification.Name("RectangleUpdatedNotification")
+    static let shpaeUpdated = Notification.Name("ShapeUpdatedNotification")
     static let planeUpdated = Notification.Name("PlaneUpdatedNotification")
 }
